signed/unsigned comparison fix.

+++ engine/minskyTensorOps.cc
@@ -325,7 +325,7 @@ namespace minsky
       if (rank()==0) return;
       
       auto& xv=hc.xvectors[rank()==1? 0: dimension];
-      if (abs(delta)>=xv.size())
+      if (size_t(abs(delta))>=xv.size())
         throw std::runtime_error("Δ ("+to_string(abs(delta))+") larger than dimension size ("+to_string(xv.size())+")");
       if (delta>=0)
         xv.erase(xv.begin(), xv.begin()+delta);
